add configure flags --optimization, and --faust-float, switch to double precision as default, add faust version '0.9.23','0.9.24' to good_faust_versions

+++ trunk/src/faust-cc/AntiAlias.cc
@@ -2,7 +2,7 @@ namespace AntiAlias {
 // generated from file '../src/faust/AntiAlias.dsp'
 
 FAUSTFLOAT 	fslider0;
-float 	fRec0[4];
+double 	fRec0[4];
 int	fSamplingFreq;
 
 void init(int samplingFreq)
@@ -13,9 +13,9 @@ void init(int samplingFreq)
 
 void compute(int count, float *input0, float *output0)
 {
-	float 	fSlow0 = fslider0;
+	double 	fSlow0 = fslider0;
 	for (int i=0; i<count; i++) {
-		fRec0[0] = (1e-20f + ((float)input0[i] + (fSlow0 * fRec0[3])));
+		fRec0[0] = (1e-20 + ((double)input0[i] + (fSlow0 * fRec0[3])));
 		output0[i] = (FAUSTFLOAT)fRec0[0];
 		// post processing
 		for (int i=3; i>0; i--) fRec0[i] = fRec0[i-1];
@@ -25,7 +25,7 @@ void compute(int count, float *input0, float *output0)
 static struct RegisterParams { RegisterParams(); } RegisterParams;
 RegisterParams::RegisterParams()
 {
-	registerVar("anti_aliase.feedback","Feedback","S","",&fslider0, 0.3f, 0.3f, 0.9f, 0.01f);
+	registerVar("anti_aliase.feedback","Feedback","S","",&fslider0, 0.3, 0.3, 0.9, 0.01);
 	registerInit(init);
 }
